requirements.txt
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
commitizen==3.27.0
decli==0.6.2
deprecation==2.1.0
exceptiongroup==1.2.1
frozenlist==1.4.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx>=0.24,<0.26
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
postgrest==0.16.8
prompt-toolkit==3.0.36
pydantic==2.7.3
pydantic_core==2.18.4
pydentic==0.0.1.dev3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.20
PyYAML==6.0.1
questionary==2.0.1
realtime==1.0.5
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
storage3==0.7.6
StrEnum==0.4.15
supabase-py-async==2.5.6
supafunc==0.4.0
termcolor==2.4.0
tomlkit==0.12.5
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.19.2
